+++ tools/Azure.Mcp.Tools.Authorization/tests/Azure.Mcp.Tools.Authorization.UnitTests/RoleAssignmentListCommandTests.cs
@@ -66,10 +66,12 @@ public class RoleAssignmentListCommandTests
                 Condition = "ActionMatches{'Microsoft.Authorization/roleAssignments/write'}"
             }
         };
-        _authorizationService.ListRoleAssignments(
+        _authorizationService.ListRoleAssignmentsAsync(
+                Arg.Is(subscriptionId),
                 Arg.Is(scope),
                 Arg.Any<string>(),
-                Arg.Any<RetryPolicyOptions>())
+                Arg.Any<RetryPolicyOptions>(),
+                Arg.Any<CancellationToken>())
             .Returns(expectedRoleAssignments);
         var command = new RoleAssignmentListCommand(_logger);
         var args = command.GetCommand().Parse([
@@ -98,7 +100,7 @@ public class RoleAssignmentListCommandTests
         // Arrange
         var subscriptionId = "00000000-0000-0000-0000-000000000001";
         var scope = $"/subscriptions/{subscriptionId}/resourceGroups/rg1";
-        _authorizationService.ListRoleAssignments(scope, null, null)
+        _authorizationService.ListRoleAssignmentsAsync(subscriptionId, scope, null, null, CancellationToken.None)
             .Returns([]);
 
         var command = new RoleAssignmentListCommand(_logger);
@@ -130,7 +132,7 @@ public class RoleAssignmentListCommandTests
         var subscriptionId = "00000000-0000-0000-0000-000000000001";
         var scope = $"/subscriptions/{subscriptionId}/resourceGroups/rg1";
 
-        _authorizationService.ListRoleAssignments(scope, null, Arg.Any<RetryPolicyOptions>())
+        _authorizationService.ListRoleAssignmentsAsync(subscriptionId, scope, null, null, CancellationToken.None)
             .ThrowsAsync(new Exception(expectedError));
 
         var command = new RoleAssignmentListCommand(_logger);